Move expand button close to its parent
